--- Slider/Assets/OceanControllerSupportEnabler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 458f1dcf7c91eae44b08e6fa70fa29c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
