+++ Slider/Assets/Scripts/UI/Controller Support.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 94eb2fdb35e7508438754149e9dd9739
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
