+++ Slider/Assets/Scripts/UI/Controller Support/OceanControllerSupportButtonsHolder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58a3c7642cee1244884e8817fbeba0a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
