--- Slider/Assets/Scripts/UI/OceanControllerSupportButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 58870e91e3bdfbd40bc77a23f3f35e8d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
